Moves command tree construction to Driver
